Math Fix & Monitor Support
Math can be hard at times.  The reactor will now generate power from anything under .35 to .65 of reactors capacity.  Above .65 and it will cease generating power.  Additionally, I've added some code to display to a monitor that is attached to the computer controlling the reactor.  If you have a monitor connected and wish to use it, remove the -- from each monitor line.